created views and forms modules
